--- src/app/(navbar)/my/_types/my.types.ts
@@ -1,9 +1,0 @@
-export interface ListItemProps {
-  title: string;
-  link?: string;
-  categories?: string[];
-  date?: string;
-  theme?: "menu" | "notice" | "faq";
-  iconSrc?: string;
-  onClickFn?: (e: React.MouseEvent<HTMLButtonElement>) => void;
-}
