+++ src/app/types.ts
@@ -10,3 +10,14 @@ export interface HashtagItem {
   searchedCount: number;
   rank: number;
 }
+
+// mypage
+export interface ListItemProps {
+  title: string;
+  link?: string;
+  categories?: string[];
+  date?: string;
+  theme?: "menu" | "notice" | "faq";
+  iconSrc?: string;
+  onClickFn?: (e: React.MouseEvent<HTMLButtonElement>) => void;
+}
